get up emprendedor2.rb

+++ emprendedor1.rb
@@ -8,7 +8,7 @@
 # Ingrese el precio, el número de usuarios y los gastos. El programa calcula las utilidades.
 precio_venta = ARGV[0].to_f
 usuarios = ARGV[1].to_i 
-gastos = ARGV[2].to_i
+gastos = ARGV[2].to_f
 utilidad =precio_venta*usuarios-gastos
 if utilidad >= 0
     utilidad =(precio_venta*usuarios-gastos)-utilidad*0.35
